pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomy-engine"
version = "1.3.0"
description = "Prefect-based autonomous build engine with human-in-the-loop decision gates"
requires-python = ">=3.10"
dependencies = [
    "prefect>=3.0",
    "anthropic",
    "openai",
    "pydantic>=2.11",
    "python-dotenv",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]
dashboard = [
    "streamlit>=1.30",
    "plotly>=5.0",
]

[tool.setuptools.packages.find]
include = ["engine*", "flows*", "intake*", "tasks*", "dashboard*"]

[tool.ruff]
line-length = 100
target-version = "py310"
[tool.ruff.lint.per-file-ignores]
# load_dotenv() must run before other imports
"flows/autonomous_flow.py" = ["E402"]
"intake/intake.py" = ["E402"]
"bench/benchmark_runs.py" = ["E402"]
"engine/llm_provider.py" = ["E402"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["slow: tests that create real virtualenvs (~30s each)"]
